Added LoginForm,RegForm, as individual components.  Added css style pages for each component.  Currently, AuthPage only loads LoginForm.  Working on a loaderbutton for the RegForm(spinnner).  This would use another package dependency(Font Awesome)
