+++ client/src/pages/AuthPage/AuthPage.js
@@ -1,47 +1,13 @@
-import React, { useState } from 'react'
-import {Button, FormGroup, FormControl, FormLabel} from 'react-bootstrap'
-//import "./Login.css";
+import React from 'react'
+import "./LoginForm.css";
+import "./RegForm.css";
+import Login from '../../components/LoginAndReg/Login';
 
 
 function AuthPage(props) {
-        const [email, setEmail] = useState("");
-        const [password, setPassword] = useState("");
-
-        
-        function validateForm() {
-            return email.length > 0 && password.length > 0;
-        }
-
-        function handleSubmit(event) {
-            event.preventDefault();
-        }
-
         return (
-            <div className="AuthPage">
-                <form onSubmit={handleSubmit}>
-                    <FormGroup controlId="email" bsSize="large">
-                        <FormLabel>Email</FormLabel>
-                        <FormControl
-                        autoFocus
-                        type="email"
-                        value={email}
-                        onChange={e => setEmail(e.target.value)}
-                        />
-                    </FormGroup>
-                    <FormGroup controlId="password" bsSize="large">
-                        <FormLabel>Password</FormLabel>
-                        <FormControl
-                        value={password}
-                        onChange={e => setPassword(e.target.value)}
-                        type="password"
-                        />
-                    </FormGroup>
-                    <Button block bsSize="large" disabled={!validateForm()} type="submit">
-                        Login
-                    </Button>
-                </form>
-            </div>
-        )
+            <Login/>
+        ) 
 }
 
 export default AuthPage
